requirements.txt
streamlit==1.15.0
requests==2.28.1
flask==2.1.1
werkzeug==2.0.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.50.0
